communicating with server to get the randomly chosen answer word
The matching words now come from more parts of the wikipedia page (not just a tags but li, p, and span tags). The javascript also sends a get request to a python server that generates a new word every 10 minutes from the most 1000 common words. Some style changes too and the old javascript is there as a separate file