retaining previous guesses clues
keeps displaying previous connections and what the clue was.
added functionality so that you can click on a connection and it will search that.
removed conncetions that were lists, main text, and subtitles, so that only link pages are used as connections (makes it neater and a little less information overloadig)